algunos cambios en la funcion fetch que ahora pasa a llamarse ToDoRequest
