+++ go.mod
@@ -2,4 +2,4 @@ module github.com/mvitta/GoRoutineMiniProject
 
 go 1.22.1
 
-require github.com/joho/godotenv v1.5.1 // indirect
+require github.com/joho/godotenv v1.5.1 // direct
